requirements.txt
aioredis==2.0.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
click==8.1.3
Django==4.2
django-otp==1.1.6
fastapi==0.95.1
fastapi-limiter==0.1.5
gunicorn==20.1.0
h11==0.14.0
idna==3.4
install==1.3.5
-e local_extension
mysqlclient==2.1.1
pydantic==1.10.7
python-dotenv==1.0.0
redis==4.5.4
Secweb==1.6.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
